ref: Split tests, fix types
