+++ mypy.ini
@@ -97,7 +97,7 @@ files = fixtures/mypy-stubs,
         src/sentry/runner/commands/devservices.py,
         src/sentry/runner/initializer.py,
         src/sentry/search/base.py,
-        src/sentry/search/events/builder.py,
+        src/sentry/search/events/builder/,
         src/sentry/search/events/constants.py,
         src/sentry/search/events/types.py,
         src/sentry/search/events/datasets/base.py,
